LayoutTests/fast/events/message-channel-listener-circular-ownership-expected.txt

CONSOLE MESSAGE: line 19: ReferenceError: Can't find variable: MessageChannel
Test that a leak is not created by assigning a MessagePort to a property of its own listener.

Only works with run-webkit-tests --leaks.
